Fix review comments
Fix comments
