+++ backends/samsung/_passes/fold_qdq.py
@@ -32,4 +32,5 @@ class FoldQDQPass(ExportPass):
         self._fold(graph_module)
         graph_module.recompile()
         dead_code_elimination_pass(graph_module)
+        _ = super().call(graph_module).graph_module
         return PassResult(graph_module, True)
