+++ backends/samsung/_passes/remove_useless_ops.py
@@ -24,6 +24,41 @@ class RemoveUselessOpPass(ExportPass):
     def __init__(self):
         super().__init__()
 
+    def gen_pattern_as_strided_copy(self, graph_module: GraphModule):
+        for node in list(graph_module.graph.nodes):  # noqa: C416
+            if node.target != exir_ops.edge.aten.mean.dim:
+                continue
+            if len(node.users) != 1:
+                continue
+            successor = list(node.users.keys())[0]
+            if successor.target != exir_ops.edge.aten.as_strided_copy.default:
+                continue
+            is_pattern = True
+            count = 0
+            for i, stride in enumerate(successor.args[2]):
+                if stride < node.meta["val"].size()[i]:
+                    if stride == 1:
+                        count += 1
+                    else:
+                        is_pattern = False
+                        break
+                if count >= 2:
+                    is_pattern = False
+                    break
+            if is_pattern:
+                yield successor
+
+    def _fold_as_strided_copy(
+        self,
+        graph_module: GraphModule,
+    ):
+        for as_strided_copy_node in self.gen_pattern_as_strided_copy(graph_module):
+            for user in list(as_strided_copy_node.users.keys()):
+                user.replace_input_with(
+                    as_strided_copy_node, as_strided_copy_node.args[0]
+                )
+            graph_module.graph.erase_node(as_strided_copy_node)
+
     def _remove_useless(
         self,
         graph_module: GraphModule,
@@ -42,9 +77,11 @@ class RemoveUselessOpPass(ExportPass):
             for user in [user for user in node.users.keys()]:  # noqa: C416
                 user.replace_input_with(node, node.all_input_nodes[0])
             graph_module.graph.erase_node(node)
+        self._fold_as_strided_copy(graph_module)
 
     def call(self, graph_module: GraphModule):
         self._remove_useless(graph_module)
         graph_module.recompile()
         dead_code_elimination_pass(graph_module)
+        _ = super().call(graph_module).graph_module
         return PassResult(graph_module, True)
